NP-1791 Move useEffect to own component
