+++ src/App.tsx
@@ -4,25 +4,22 @@ import { useDispatch, useSelector } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import styled from 'styled-components';
 import { CircularProgress } from '@material-ui/core';
-import { AuthorityQualifiers, addQualifierIdForAuthority, getAuthority } from './api/authorityApi';
 import { getCurrentUserAttributes } from './api/userApi';
 import Footer from './layout/Footer';
 import Header from './layout/header/Header';
 import Notifier from './layout/Notifier';
 import AuthorityOrcidModal from './pages/user/authority/AuthorityOrcidModal';
-import { setAuthorityData, setPossibleAuthorities, setUser, setRoles } from './redux/actions/userActions';
+import { setUser, setRoles } from './redux/actions/userActions';
 import { RootStore } from './redux/reducers/rootReducer';
 import { awsConfig } from './utils/aws-config';
 import { USE_MOCK_DATA } from './utils/constants';
 import { mockUser } from './utils/testfiles/mock_feide_user';
 import AppRoutes from './AppRoutes';
-import useFetchAuthorities from './utils/hooks/useFetchAuthorities';
 import { setNotification } from './redux/actions/notificationActions';
 import { getInstitutionUser } from './api/roleApi';
 import { NotificationVariant } from './types/notification.types';
 import { InstitutionUser } from './types/user.types';
-import { useTranslation } from 'react-i18next';
-import { Authority } from './types/authority.types';
+import useFetchCurrentAuthority from './utils/hooks/useFetchCurrentAuthority';
 
 const StyledApp = styled.div`
   min-height: 100vh;
@@ -49,10 +46,9 @@ const ProgressContainer = styled.div`
 
 const App: FC = () => {
   const dispatch = useDispatch();
-  const { t } = useTranslation('feedback');
   const user = useSelector((store: RootStore) => store.user);
   const [isLoadingUser, setIsLoadingUser] = useState(true);
-  const [matchingAuthorities, isLoadingMatchingAuthorities] = useFetchAuthorities(user?.name ?? '');
+  useFetchCurrentAuthority();
 
   useEffect(() => {
     // Setup aws-amplify
@@ -107,41 +103,6 @@ const App: FC = () => {
     }
   }, [dispatch, user]);
 
-  useEffect(() => {
-    if (matchingAuthorities && user && !user.authority && !user.possibleAuthorities) {
-      const fetchAuthority = async () => {
-        const filteredAuthorities = matchingAuthorities.filter((auth) => auth.feideids.some((id) => id === user.id));
-        if (filteredAuthorities.length === 1) {
-          // Use exsisting authority
-          const existingScn = filteredAuthorities[0].systemControlNumber;
-          const existingAuthority = await getAuthority(existingScn);
-          if (existingAuthority?.error) {
-            dispatch(setNotification(t('error.get_authority'), NotificationVariant.Error));
-          } else if (existingAuthority?.data) {
-            let currentAuthority = existingAuthority.data;
-            if (user.cristinId && !existingAuthority.data.orgunitids.includes(user.cristinId)) {
-              // Add cristinId to Authority's orgunitids
-              const authorityWithOrgId = await addQualifierIdForAuthority(
-                existingScn,
-                AuthorityQualifiers.ORGUNIT_ID,
-                user.cristinId
-              );
-              if (authorityWithOrgId?.error) {
-                dispatch(setNotification(authorityWithOrgId.error, NotificationVariant.Error));
-              } else {
-                currentAuthority = authorityWithOrgId as Authority;
-              }
-            }
-            dispatch(setAuthorityData(currentAuthority));
-          }
-        } else {
-          dispatch(setPossibleAuthorities(matchingAuthorities));
-        }
-      };
-      fetchAuthority();
-    }
-  }, [dispatch, t, matchingAuthorities, user]);
-
   return isLoadingUser ? (
     <ProgressContainer>
       <CircularProgress />
@@ -156,7 +117,7 @@ const App: FC = () => {
         </StyledContent>
         <Footer />
       </StyledApp>
-      {user && !isLoadingMatchingAuthorities && (user.authority || user.possibleAuthorities) && <AuthorityOrcidModal />}
+      {user && (user.authority || user.possibleAuthorities) && <AuthorityOrcidModal />}
     </BrowserRouter>
   );
 };
